pyproject.toml
[project]
name = "energy-intel"
version = "0.1.0"
description = "Permian-focused daily energy intelligence brief"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2",
    "requests>=2.32",
    "yfinance>=0.2.40",
    "fredapi>=0.5.2",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
